+++ cli/cli_utils.go
@@ -68,8 +68,8 @@ func valClients(allClients clients.OrderedClients, flags *GenCmdFlags, services 
 			executionClient.Name = executionParts[0]
 			if len(executionParts) > 1 {
 				log.Warn(configs.CustomExecutionImagesWarning)
-				executionClient.Image = strings.Join(executionParts[1:], ":")
 			}
+			executionClient.SetImageOrDefault(strings.Join(executionParts[1:], ":"))
 		}
 		if err = clients.ValidateClient(executionClient, execution); err != nil {
 			return nil, err
@@ -88,8 +88,8 @@ func valClients(allClients clients.OrderedClients, flags *GenCmdFlags, services 
 			consensusClient.Name = consensusParts[0]
 			if len(consensusParts) > 1 {
 				log.Warn(configs.CustomConsensusImagesWarning)
-				consensusClient.Image = strings.Join(consensusParts[1:], ":")
 			}
+			consensusClient.SetImageOrDefault(strings.Join(consensusParts[1:], ":"))
 		}
 		if err = clients.ValidateClient(consensusClient, consensus); err != nil {
 			return nil, err
@@ -108,9 +108,8 @@ func valClients(allClients clients.OrderedClients, flags *GenCmdFlags, services 
 			validatorClient.Name = validatorParts[0]
 			if len(validatorParts) > 1 {
 				log.Warn(configs.CustomValidatorImagesWarning)
-				validatorClient.Image = strings.Join(validatorParts[1:], ":")
-
 			}
+			validatorClient.SetImageOrDefault(strings.Join(validatorParts[1:], ":"))
 		}
 		if err = clients.ValidateClient(validatorClient, validator); err != nil {
 			return nil, err