+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support the new command `run`, used to run all the services generated.
 - Support for PPA packaging (apt install).
 - Added flag `--container-tag` to add a suffix to sedge containers name.
+- Support the new command `import-key` to import validator keys in an existing configuration.
 
 ### Changed